refactor: rename TableBranch to TableVersioning
Rename components to better reflect their scope (branches, tags, snapshots):
- TableBranch.vue → TableVersioning.vue
- TableBranchVisualization.vue → TableVersioningVisualization.vue
- Update all imports and exports in index.ts
- Update CLAUDE.md documentation
