+++ src/components/TableVersioningVisualization.vue
